+++ shaders/plot_dot.geom
@@ -3,8 +3,8 @@
 layout(location =  0) uniform mat4    pmv;
 layout(location =  1) uniform double  t_min;
 layout(location =  2) uniform double  t_max;
-layout(location =  3) uniform double  v_min;
-layout(location =  4) uniform double  v_max;
+layout(location =  3) uniform float   v_min;
+layout(location =  4) uniform float   v_max;
 layout(location =  5) uniform double  width;
 layout(location =  6) uniform double  height;
 layout(location =  7) uniform float   y_offset;
@@ -30,9 +30,9 @@ void main()
 
     float x, y;
     double r_t = t_max-t_min;
-    double r_v = v_max-v_min;
+    float r_v = v_max - v_min;
     x = float(width  *       (gs_in[0].t-t_min)/r_t  );
-    y = float(height * (1.lf-(gs_in[0].v-v_min)/r_v) ) + y_offset;
+    y = float(height * (1.lf - (gs_in[0].v - v_min) / r_v)) + y_offset;
     gl_Position = pmv * vec4(x, y, 0, 1);
     gl_PointSize = 5.0;
     EmitVertex();
